修改热更新触发时机
